backend/requirements.txt
flask
flask-cors
spotipy
openai
requests
textblob
python-dotenv
